Consolidate fixture namespaces
Having Schema and Fixture be separate felt off as both namespaces are
for fixtures.
