+++ tests/TestCase/TestSuite/Fixture/SchemaCleanerTest.php
@@ -13,11 +13,11 @@ declare(strict_types=1);
  * @since         4.3.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Test\TestCase\TestSuite\Schema;
+namespace Cake\Test\TestCase\TestSuite\Fixture;
 
 use Cake\Database\Schema\TableSchema;
 use Cake\Datasource\ConnectionManager;
-use Cake\TestSuite\Schema\SchemaCleaner;
+use Cake\TestSuite\Fixture\SchemaCleaner;
 use Cake\TestSuite\TestCase;
 
 class SchemaCleanerTest extends TestCase
